# docs/swagger.yaml
definitions:
...
  ds.ImpurityFractionExperiment:
    properties:
      created_at:
        type: string
      creator:
        $ref: '#/definitions/ds.User'
      creator_id:
        type: string
      finished_at:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      moderator:
        $ref: '#/definitions/ds.User'
      moderator_id:
        type: string
      molar_volume:
        type: number
      status:
        type: string
    type: object
  ds.User:
    properties:
      id:
        type: string
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    type: object
...
  dto.DraftExperimentResponse:
    properties:
      experiment_id:
        type: integer
      sample_count:
        type: integer
    type: object
  dto.ExperimentResponse:
    properties:
      creator_login:
        type: string
      experiment:
        $ref: '#/definitions/ds.ImpurityFractionExperiment'
      experiment_sample_cards:
        items:
          $ref: '#/definitions/dto.ExperimentSampleCard'
        type: array
      moderator_login:
        type: string
    type: object
  dto.ExperimentSampleCard:
    properties:
      evolvedGasVolume:
        type: string
      formula:
        type: string
      imageURL:
        type: string
      massFractionPercentage:
        type: string
      relativeMolecularMass:
        format: float64
        type: number
      sampleID:
        type: integer
      sampleMass:
        type: string
      stoichiometricCoefficient:
        format: float64
        type: number
      title:
        type: string
    type: object
  dto.ExperimentsResponse:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      finished_at:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      moderator_login:
        type: string
      molar_volume:
        type: number
      status:
        type: string
    type: object
  dto.FormExperiment:
    properties:
      creator_login:
        type: string
      experiment:
        $ref: '#/definitions/ds.ImpurityFractionExperiment'
    type: object
  dto.ModerateExperiment:
    properties:
      creator_login:
        type: string
      experiment:
        $ref: '#/definitions/ds.ImpurityFractionExperiment'
      moderator_login:
        type: string
    type: object
  dto.SoftDeleteExperiment:
    properties:
      creator_login:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      status:
        type: string
    type: object
  dto.StatusJSON:
    properties:
      status:
        type: string
    type: object
  dto.UpdateExperiment:
    properties:
      creator_login:
        type: string
      experiment:
        $ref: '#/definitions/ds.ImpurityFractionExperiment'
    type: object
...
info:
  contact: {}
paths:
  /impurity-experiments:
    get:
      description: Возвращает исследования с возможностью фильтрации по датам и статусу
      parameters:
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: from-date
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: to-date
        type: string
      - description: Статус исследования
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список исследований
          schema:
            items:
              $ref: '#/definitions/dto.ExperimentsResponse'
            type: array
        "400":
          description: Неверный формат даты
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список исследований
      tags:
      - impurity-experiments
  /impurity-experiments/{id}:
    delete:
      description: Выполняет логическое удаление исследования
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус удаления
          schema:
            $ref: '#/definitions/dto.SoftDeleteExperiment'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить исследование
      tags:
      - impurity-experiments
    get:
      description: Возвращает полную информацию об исследовании включая образцы
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные исследования с образцами
          schema:
            $ref: '#/definitions/dto.ExperimentResponse'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить исследование по ID
      tags:
      - impurity-experiments
    put:
      consumes:
      - application/json
      description: Обновляет данные исследования
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные исследования
        in: body
        name: experiment
        required: true
        schema:
          $ref: '#/definitions/ds.ImpurityFractionExperiment'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленное исследование
          schema:
            $ref: '#/definitions/dto.UpdateExperiment'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить исследование
      tags:
      - impurity-experiments
  /impurity-experiments/{id}/form:
    put:
      description: Переводит исследование в статус "formed"
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сформированное исследование
          schema:
            $ref: '#/definitions/dto.FormExperiment'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать исследование
      tags:
      - impurity-experiments
  /impurity-experiments/{id}/moderation:
    put:
      consumes:
      - application/json
      description: Изменяет статус исследования (только для модераторов)
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.StatusJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Результат модерации
          schema:
            $ref: '#/definitions/dto.ModerateExperiment'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Модерировать исследование
      tags:
      - impurity-experiments
...
  /impurity-experiments/draft:
    get:
      description: Возвращает информацию о текущем черновике исследования пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Данные корзины исследования
          schema:
            $ref: '#/definitions/dto.DraftExperimentResponse'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить корзину исследования
      tags:
      - impurity-experiments
...
swagger: "2.0"